img-classifier-hub-on/function/requirements.txt
# tensorflow==2.3.0 # temporal until available wheels
tensorflow-hub
pillow
pybind11
Cython==0.29.21
h5py==2.10.0
grpcio
scipy
numpy

